safari-600.7-branch/LayoutTests/sputnik/Conformance/07_Lexical_Conventions/7.9_Automatic_Semicolon_Insertion/S7.9_A6.4_T2-expected.txt

CONSOLE MESSAGE: line 77: SyntaxError: Unexpected token ';'
S7.9_A6.4_T2

PASS Expected parsing failure

TEST COMPLETE
